chore(docus): add blobity composable
