--- docs2/components/Foo.vue
@@ -1,3 +1,0 @@
-<template>
-  Hello World
-</template>
